requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Audio Processing
whisper==1.1.10
pyannote.audio==3.1.1
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1

# NLP and ML
transformers==4.35.2
torch==2.1.1
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 